api/GCP/requirements.txt
tensorflow
google-cloud-storage
Pillow
numpy
